Editorial: Correct references to "result type"
"Commands" are generally described to include a "result type," but none
of the commands defined actually include such a value. They do include a
"return type," as defined by WEBIDL. Since WEBIDL's "return type" does
not seem particularly relevant and since the terms "result type" and
"return type" have similar connotations in this context, this appears to
be an editorial oversight.

Update the command definitions to specify a "result type," and update
the explanation of the commands' "remote end steps" to return an
instance of the "result type."

See also the corresponding change to the WebDriver BiDi proposal:
https://github.com/w3c/webdriver-bidi/pull/192