Editorial: rename "respond with an error"
The proposal defines "error" as an internal type for describing one kind
of result of fallible algorithms, and it defines "ErrorResponse" as a
type of message which may be sent to clients. The algorithm used to
transmit error responses to clients is named "respond with an error"
even though it does not use the "error" type.

Rename "respond with an error" to "send an error response" to more
accurately describe the types involved.