Normative: destroy sessions with zero connections
The algorithm for the "session.new" command includes a guard for the
case where an active session already exists:

> 2. If the list of [=active sessions=] is not empty, then return
>    [=error=] with error code [=session not created=].

Prior to this commit, sessions were not removed from this list when they
were depleted of connections. Following the creation and termination of
some initial connection, a remote end would forever deny further
requests to create new sessions.

Destroy the active session when its final connection is closed so that a
new session may be subsequently created.