Normative: specify settings commands (#54)
Prior to this commit, the behavior of all three commands in the
"<vendor>:settings" module was entirely implementation-defined. Although
it is known that today's implementations require some leeway in their
support of this behavior, the high degree of ambiguity obfuscated even
the failure modes. This would leave clients and implementations guessing
about when and how errors might be reported. Additionally, the prose did
not communicate the rationale for the ambiguity.

Specify the desired behavior while leaving implementations the
flexibility they are currently understood to require. Document the
rationale for this flexibility, including the editorial expectation
regarding its eventual removal.