Rewrite explanation of "generic AT driver" soln.
This section previously described an alternative to standardization and
explained how we are currently exploring that alternative. By neglecting
to highlight the alternative's deficiencies (or explain our limited
expectations of our experiment), it risked confusing readers about how
it fits in the standardization process.

Rewrite the section to more clearly describe this as a fundamentally
insufficient solution while still documenting the active research behind
it.