Refactor "pressKeys" as one of many "user intents" (#76)
In conversation with assistive technology vendors, we have learned that
automating arbitrary key presses does not necessarily align with the
security model of all platforms. Separately, we anticipate introducing
means to simulate more specific types of user interaction (e.g. "move to
next heading"). Because these may themselves be implemented at the
vendor's discretion (and because we wish to allow for vendors to
experiment with the development of still more kinds of interaction), the
space for simulating interaction may be somewhat fragmented,
particularly in the initial deployments of this protocol. We have
decided to design a command which accommodates this variability so that
the presence/absence of interactions can be communicated consistently
(note the "unknown user intent" error) and so that additional
interactions can be introduced without the addition of entirely new
commands (or entirely new extension commands, as the case may be).