Re-order CDDL definitions so that Command / Message are first
The first rule in a CDDL document defines the entry point to validate a CBOR
data item. The CDDL of each module started with `Extensible`, which essentially
made any CBOR data item valid.

This update makes `Command` and `Message` the first CDDL rules in each module
(as done in WebDriver BiDi), and defines `Extensible` is a single block common
to both the remote end and local end definitions.