Adapt CDDL declarations to new version of Bikeshed
This updates the CDDL declarations to leverage new support for CDDL
highlighting and linking in Bikeshed, see details in:
https://github.com/w3c/webdriver-bidi/pull/927

Note in particular the need to use `{^ ^}` to wrap references to the CDDL
module definitions, and the use of a dedicated `data-cddl-module` attribute to
associate a CDDL block with the right module.

Bikeshed automatically generates a CDDL index (per module) as an annex.